progettobot/requirements.txt
python-telegram-bot requests
python-telegram-bot
# googlesearch-python
google-search-client
